Resolve #37 (#41)
* Ability to order a domain with SBD or STEEM payment

* More domains validations. Supported TLDs list added. Docker secrets

* One more DNS record added for newly created domain

* Account switch

+++ request_wildcard_certificate.sh
@@ -1,10 +1,14 @@
-docker run -it --rm --name letsencrypt \
-    -v "engrave_nginx_certs:/etc/letsencrypt" \
-    -v "engrave_nginx_certs:/var/lib/letsencrypt" \
-    quay.io/letsencrypt/letsencrypt:latest \
-        certonly \
-        -d engrave.website \
-        -d *.engrave.website \
-        --manual \
-        --preferred-challenges dns \
-        --server https://acme-v02.api.letsencrypt.org/directory
+if [ "$1" != "" ]; then
+    docker run -it --rm --name letsencrypt \
+        -v "engrave_nginx_certs:/etc/letsencrypt" \
+        -v "engrave_nginx_certs:/var/lib/letsencrypt" \
+        quay.io/letsencrypt/letsencrypt:latest \
+            certonly \
+            -d $1 \
+            -d *.$1 \
+            --manual \
+            --preferred-challenges dns \
+            --server https://acme-v02.api.letsencrypt.org/directory
+else 
+    echo "You need to provide domain name!"
+fi
